fix: make title bar cross-platform (macOS-compatible)
- Revert decorations to true in tauri.conf.json; disable at runtime on
  non-macOS via setDecorations(false)
- Conditionally render custom window controls (non-macOS only)
- Fix resetZoom to use zoomTo(1) instead of setViewport
- Normalize animation durations to 300ms
- Extract shared layout actions to reduce duplication
- Revert cosmetic formatting changes in PanelLayout and lib.rs
