+++ src-tauri/capabilities/default.json
@@ -10,6 +10,7 @@
     "core:window:allow-destroy",
     "core:window:allow-minimize",
     "core:window:allow-toggle-maximize",
+    "core:window:allow-set-decorations",
     "core:window:allow-start-dragging",
     "dialog:allow-open",
     "dialog:allow-save",
