+++ src-tauri/tauri.conf.json
@@ -19,7 +19,7 @@
         "minHeight": 768,
         "resizable": true,
         "fullscreen": false,
-        "decorations": false,
+        "decorations": true,
         "dragDropEnabled": false
       }
     ],
