+++ src/components/layout/ProjectTitleBar.tsx
@@ -9,6 +9,7 @@ import { useUIStore } from "@/stores/uiStore";
 import { useSettingsStore } from "@/stores/settingsStore";
 import { resolveKeybinding } from "@/config/keybindings";
 import { exportAssetPack, exportCurrentJson } from "@/utils/exportAssetPack";
+import { isMac } from "@/utils/platform";
 
 interface MenuItemProps {
   label: string;
@@ -105,11 +106,11 @@ export function ProjectTitleBar({
   const selectedCount = useEditorStore((s) => s.nodes.filter((n) => n.selected).length);
 
   function handleResetZoom() {
-    reactFlow.setViewport({ x: 0, y: 0, zoom: 1 }, { duration: 200 });
+    reactFlow.zoomTo(1, { duration: 300 });
   }
 
   function handleFitView() {
-    reactFlow.fitView({ padding: 0.1, duration: 200 });
+    reactFlow.fitView({ padding: 0.1, duration: 300 });
   }
 
   function handleZoomToSelection() {
@@ -118,11 +119,13 @@ export function ProjectTitleBar({
     reactFlow.fitView({
       nodes: nodes.map((n) => ({ id: n.id })),
       padding: 0.2,
-      duration: 200,
+      duration: 300,
     });
   }
 
   useEffect(() => {
+    if (isMac) return;
+
     appWindow.isMaximized().then(setIsMaximized);
 
     const unlisten = appWindow.onResized(() => {
@@ -148,7 +151,7 @@ export function ProjectTitleBar({
 
   return (
     <div
-      data-tauri-drag-region
+      {...(!isMac ? { "data-tauri-drag-region": true } : {})}
       className="h-8 bg-tn-panel border-b border-tn-border flex items-center justify-between select-none shrink-0"
     >
       <div className="flex items-center">
@@ -312,33 +315,13 @@ export function ProjectTitleBar({
         </MenuDropdown>
       </div>
 
-      <div className="flex items-center gap-0">
-        <button
-          onClick={handleMinimize}
-          className="w-12 h-8 flex items-center justify-center hover:bg-tn-surface transition-colors text-tn-text-muted hover:text-tn-text"
-          title="Minimize"
-        >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="16"
-            height="16"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
+      {!isMac && (
+        <div className="flex items-center gap-0">
+          <button
+            onClick={handleMinimize}
+            className="w-12 h-8 flex items-center justify-center hover:bg-tn-surface transition-colors text-tn-text-muted hover:text-tn-text"
+            title="Minimize"
           >
-            <line x1="5" y1="12" x2="19" y2="12" />
-          </svg>
-        </button>
-
-        <button
-          onClick={handleMaximize}
-          className="w-12 h-8 flex items-center justify-center hover:bg-tn-surface transition-colors text-tn-text-muted hover:text-tn-text"
-          title={isMaximized ? "Restore" : "Maximize"}
-        >
-          {isMaximized ? (
             <svg
               xmlns="http://www.w3.org/2000/svg"
               width="16"
@@ -350,10 +333,52 @@ export function ProjectTitleBar({
               strokeLinecap="round"
               strokeLinejoin="round"
             >
-              <rect x="3" y="3" width="18" height="18" rx="2" ry="2" />
-              <line x1="9" y1="3" x2="9" y2="21" />
+              <line x1="5" y1="12" x2="19" y2="12" />
             </svg>
-          ) : (
+          </button>
+
+          <button
+            onClick={handleMaximize}
+            className="w-12 h-8 flex items-center justify-center hover:bg-tn-surface transition-colors text-tn-text-muted hover:text-tn-text"
+            title={isMaximized ? "Restore" : "Maximize"}
+          >
+            {isMaximized ? (
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                width="16"
+                height="16"
+                viewBox="0 0 24 24"
+                fill="none"
+                stroke="currentColor"
+                strokeWidth="2"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+              >
+                <rect x="3" y="3" width="18" height="18" rx="2" ry="2" />
+                <line x1="9" y1="3" x2="9" y2="21" />
+              </svg>
+            ) : (
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                width="16"
+                height="16"
+                viewBox="0 0 24 24"
+                fill="none"
+                stroke="currentColor"
+                strokeWidth="2"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+              >
+                <rect x="3" y="3" width="18" height="18" rx="2" ry="2" />
+              </svg>
+            )}
+          </button>
+
+          <button
+            onClick={handleClose}
+            className="w-12 h-8 flex items-center justify-center hover:bg-red-600 transition-colors text-tn-text-muted hover:text-white"
+            title="Close"
+          >
             <svg
               xmlns="http://www.w3.org/2000/svg"
               width="16"
@@ -365,32 +390,12 @@ export function ProjectTitleBar({
               strokeLinecap="round"
               strokeLinejoin="round"
             >
-              <rect x="3" y="3" width="18" height="18" rx="2" ry="2" />
+              <line x1="18" y1="6" x2="6" y2="18" />
+              <line x1="6" y1="6" x2="18" y2="18" />
             </svg>
-          )}
-        </button>
-
-        <button
-          onClick={handleClose}
-          className="w-12 h-8 flex items-center justify-center hover:bg-red-600 transition-colors text-tn-text-muted hover:text-white"
-          title="Close"
-        >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="16"
-            height="16"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          >
-            <line x1="18" y1="6" x2="6" y2="18" />
-            <line x1="6" y1="6" x2="18" y2="18" />
-          </svg>
-        </button>
-      </div>
+          </button>
+        </div>
+      )}
     </div>
   );
 }
