+++ src/utils/platform.ts
@@ -1,0 +1,1 @@
+export const isMac = navigator.userAgent.includes("Mac");
